# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controller.SwaggerAttachmentListData:
    properties:
      list:
        items:
          $ref: '#/definitions/controller.SwaggerAttachmentResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  controller.SwaggerAttachmentResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      filename:
        example: 550e8400-e29b-41d4-a716-446655440000.jpg
        type: string
      height:
        example: 600
        type: integer
      id:
        example: 1
        type: integer
      mime_type:
        example: image/jpeg
        type: string
      original_name:
        example: test.jpg
        type: string
      size:
        example: 102400
        type: integer
      thumb_url:
        example: http://localhost:8080/uploads/2024/01/thumb_test.jpg
        type: string
      type:
        example: image
        type: string
      url:
        example: http://localhost:8080/uploads/2024/01/test.jpg
        type: string
      width:
        example: 800
        type: integer
    type: object
...
  utils.Response:
    properties:
      code:
        description: 状态码，0表示成功
        example: 0
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
  description: 个人博客系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
...
  /attachments:
    get:
      consumes:
      - application/json
      description: 分页获取附件列表（用户只能看到自己的文件）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 文件类型 image/file/video
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerAttachmentListData'
              type: object
      security:
      - BearerAuth: []
      summary: 获取附件列表
      tags:
      - 文件管理
  /attachments/{id}:
    delete:
      consumes:
      - application/json
      description: 删除附件（需要认证，只能删除自己的文件）
      parameters:
      - description: 附件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 附件不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除附件
      tags:
      - 文件管理
    get:
      consumes:
      - application/json
      description: 根据ID获取附件详情
      parameters:
      - description: 附件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerAttachmentResponse'
              type: object
        "404":
          description: 附件不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取附件详情
      tags:
      - 文件管理
  /attachments/my:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的所有附件（需要认证）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.SwaggerAttachmentResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户的附件
      tags:
      - 文件管理
...
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传单个文件（图片、文档等）（需要认证）
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerAttachmentResponse'
              type: object
        "400":
          description: 参数错误或文件过大
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件管理
  /upload/multiple:
    post:
      consumes:
      - multipart/form-data
      description: 同时上传多个文件（需要认证）
      parameters:
      - description: 要上传的文件列表
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功（部分失败时返回错误列表）
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    errors:
                      items:
                        type: string
                      type: array
                    success:
                      items:
                        $ref: '#/definitions/controller.SwaggerAttachmentResponse'
                      type: array
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 多文件上传
      tags:
      - 文件管理
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"